Implement MissionBuilder behavior
