+++ lib/dreadnought/core/mission/builder.ex
@@ -1,7 +1,10 @@
 defmodule Dreadnought.Core.Mission.Builder do
 
+  use Dreadnought.Core.Mission.Spec
   alias Dreadnought.Core.Mission
 
   @callback new(mission_name :: String.t) :: Mission.t
+
+  @callback mission_spec(mission_name :: String.t) :: mission_spec
 
 end
