+++ lib/dreadnought/multiplayer/new_player.ex
@@ -6,7 +6,7 @@ defmodule Dreadnought.Multiplayer.NewPlayer do
 
   use Dreadnought.Core.Mission.Spec
   alias Dreadnought.Core.Player
-  alias Dreadnought.Multiplayer.Mission, as: MissionBuilder
+  alias Dreadnought.Multiplayer.MissionBuilder
   alias Dreadnought.Util.Slugs
   alias Ecto.Changeset
 
@@ -47,11 +47,10 @@ defmodule Dreadnought.Multiplayer.NewPlayer do
   end
 
   def to_player(%__MODULE__{mission_name: mission_name} = new_player) do
-    mission_spec = new_mission_spec(MissionBuilder, mission_name)
     new_player
     |> Map.from_struct
     |> Map.drop([:mission_name])
-    |> Map.put(:mission_spec, mission_spec)
+    |> Map.put(:mission_spec, MissionBuilder.mission_spec(mission_name))
     |> Player.from_new_player
   end
